requirements.txt
pandas
numpy
pyspark
redis 
kafka-python
minio
mysql-connector-python
six